worker content moved to services
