+++ apps/dashboard/worker/index.ts
@@ -1,35 +1,38 @@
-declare const self: ServiceWorkerGlobalScope
+import '@fc/services/src/worker'
+// import '../../../packages/services/src/worker'
 
-self.addEventListener('push', event => {
-  const data = JSON.parse(event.data?.text() ?? '{ title: "" }')
-  event.waitUntil(
-    self.registration.showNotification(data.title, {
-      body: data.message,
-      icon: '/icons/android-chrome-192x192.png',
-    }),
-  )
-})
+// declare const self: ServiceWorkerGlobalScope
 
-self.addEventListener('notificationclick', event => {
-  event.notification.close()
-  event.waitUntil(
-    self.clients
-      .matchAll({ type: 'window', includeUncontrolled: true })
-      .then(clientList => {
-        if (clientList.length > 0) {
-          let client = clientList[0]
-          for (let i = 0; i < clientList.length; i++) {
-            if (clientList[i].focused) {
-              client = clientList[i]
-            }
-          }
+// self.addEventListener('push', event => {
+//   const data = JSON.parse(event.data?.text() ?? '{ title: "" }')
+//   event.waitUntil(
+//     self.registration.showNotification(data.title, {
+//       body: data.message,
+//       icon: '/icons/android-chrome-192x192.png',
+//     }),
+//   )
+// })
 
-          return client.focus()
-        }
+// self.addEventListener('notificationclick', event => {
+//   event.notification.close()
+//   event.waitUntil(
+//     self.clients
+//       .matchAll({ type: 'window', includeUncontrolled: true })
+//       .then(clientList => {
+//         if (clientList.length > 0) {
+//           let client = clientList[0]
+//           for (let i = 0; i < clientList.length; i++) {
+//             if (clientList[i].focused) {
+//               client = clientList[i]
+//             }
+//           }
 
-        return self.clients.openWindow('/')
-      }),
-  )
-})
+//           return client.focus()
+//         }
 
-export type {}
+//         return self.clients.openWindow('/')
+//       }),
+//   )
+// })
+
+// export type {}
